+++ software/app/table.h
@@ -6,7 +6,7 @@
 #include "ball.h"
 
 typedef struct {
-    uint8_t height, width;
+    uint16_t height, width;
 } table;
 
 void bounceOffTable(ball* ball);
